Skeleton structure for Parsers trait
The new traits and classes are ReaderStream, Parsers and StringParsers.

ReaderStream is a wrapper for the input (a sequence of some element). Eventually, it will be used to keep track of the position of input.

The thrust of this change is the Parsers trait. It is the entry point for the Parsers module. This module defines the abstract concepts of parsers.

The StringParsers trait extends Parsers and specifies that it parsers strings or Seq[Char].

Two combinators are defined, unit and string.