Add cats library in order to support the applicative style
Make Parser[T] an instance of the applicative typeclass.

The instance definition is in the implicit val parserApplicative. See http://eed3si9n.com/herding-cats/Applicative.html for background on Applicative and http://eed3si9n.com/herding-cats/typeclasses-102.html for how instances of typeclasses are created in cats.

By making parser an instance of Applicative, we can now use the applicative style of effectual programming. Instances of Parser[T] can be combined using the <*> or ap operator. By using this combinator, the parser "state" is strung through the calls. See the ap method implementation for how the input state is strung from one parser to the next.

Some helper implicits are added for functions wrapped in calls to pure to remove the need for calls to .curried.